WasteBot/requirements.txt
# WasteBot - Python dependencies
# Install: pip install -r requirements.txt

# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=1.0.0

# Vector store (optional):
# - On Windows, `faiss-cpu` often isn't available via `pip`.
# - This project will fall back to a local JSON vector store if FAISS isn't installed.
# faiss-cpu>=1.8.0

# Document loaders
pypdf>=4.0.0

# Embeddings
langchain-huggingface>=0.1.0
# Optional (recommended for best retrieval quality):
# sentence-transformers>=3.0.0

# LLM (Groq only)
langchain-groq>=0.1.0
groq>=0.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# API server (optional, for integrating with Backend/Frontend)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
